SPRINT 2
The second sprint of my text based adventure. I have added code for HallwayNorth, Cafe, Andy Room, and the party room, and added more items that may be added to your backpack. This is a good transition from the easier and less story based beginning to more complicated and story focused ending.